challenge_01/rogerhoward/python_pil/requirements.txt
Pillow==3.0.0
virtualenv==1.11.6
virtualenvwrapper==4.3
wsgiref==0.1.2
